albot/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "albot"
version = "0.1.0"
description = "AI Telegram bot for real estate lead generation"
authors = [{name = "AL Bot Team", email = "team@albot.ru"}]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-telegram-bot==21.6",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "httpx==0.27.2",
    "loguru==0.7.2",
    "uvloop==0.19.0",
    "google-api-python-client==2.141.0",
    "google-auth==2.34.0",
    "google-auth-oauthlib==1.2.1",
    "aiofiles==24.1.0",
    "orjson==3.10.7",
    "pandas==2.2.3",
    "openpyxl==3.1.5",
    "xlrd==2.0.1",
    "python-docx==1.1.2",
    "pdfminer.six==20240706",
    "sentry-sdk==2.39.0",
]

[project.optional-dependencies]
test = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-mock==3.12.0",
    "pytest-cov==4.1.0",
    "pytest-xdist==3.3.1",
    "psutil==5.9.6",
]

[project.urls]
Homepage = "https://github.com/albot/albot"
Documentation = "https://docs.albot.ru"
Repository = "https://github.com/albot/albot"
Issues = "https://github.com/albot/albot/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["albot*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "load: marks tests as load tests",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["albot/src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]